+++ CHANGELOG.md
@@ -12,8 +12,11 @@ code
 - Added in `end_silence_length` to track the raw end silences
 - Added new pub function `validate_input` for `VadSession` struct. `process` function will use it to make sure input is valid in debug mode.
 - Added a new method: `processed_duration`. A sample is considered as processed if it has been seen by Silero NN.
+- API to get current start/end time of session audio
+
 ### Fixed
 - Potential OOM when handling long autio.
+
 ### Changed
 - Deleted `timestamp_ms` in `SpeechEnd`.
 - Added `start_timestamp_ms`, `end_timestamp_ms`, and `samples` in `SpeechEnd`.